Create Client Library (#5)
Was getting a bit boilerplate as regards API calls, so rolled up most of
the crap into a library.  I did consider making the token management a
silent part of it, but decided against this, as unauthorized isn't just
about whether the token has expired or not, it can mean other things in
future such as Oauth2 scopes.