Cluster Provisioning Part 1
Get the basics done such as layout (a non-distracting one so you don't
get lost or tweak things you don't need to), and styling.  A quick PoC
for filling in the blanks, along with some filtering to gauge what that
looks like.  Should really hit server with some cache control...