Improve Provisioning UX
Couple improvements, in lieu of the underlaying platform "letting nova
schedule as it sees fit", a temporary fix is to just select nova by
default at the top level, and leave per pool-choices strictly optional.
The other thing is to make autoscaling "roll-up" and get implicitly
defined, rather than having to opt in.  Also adding a key pair wasn't
actually passed to the API, so set that.  Fix a heinous bug to do with
token scope checking.