Await Project Creation
The initial problem is that it creates a project, then expects it to be
immediately available, so the subsequent CP creation can actually fail.
Silently.  So first up add in a simple error notification banner that
can be dismissed.  Next up, when creating the project, block CP creation
until the project reports as provisioned, lest server reject the
creation request.