Fix Upgrade Indicator
When we select "valid" bundles for clusters and control planes, we
filter out the EoL and preview ones.  The check used to ask "am I at the
most recent version" i.e. the top of the list.  This led to a false
positive where a preview version would flag being upgradable
erroneously.  Switch to using semver and do a < check for upgradability.