Fix Binding Issues (#99)
Seems that if anything changes with an included object, all bound
objects will get an update event.  In this case it was spotted because a
flavor update was triggering all flavours to update because of an
erroneous bind, then causing an immediate reset.  Seems I'd copied and
pasted the mistake everywhere, which explains a lot of the comments
about weird behaviour!!