Fix a Couple Niggles
First up, when you add a new pool in the update menu, its autoscaling is
inconsistent and set to false.  When you try create/update a cluster
without a workload pool, it raises a vagues error.  We can catch this
before submitting, inhibit the submit, and give some useful help text.

Fixes #104
Fixes #105