Switch to local (ephemeral) disk for control plane nodes
When `.spec.rootVolume.diskSize` is set to 0, the instance uses
ephemeral storage as configured as part of the flavor instead of
creating a new volume and booting from it.

This is desirable in the case of etcd since it is particularly sensitive
to disk write latency, and thus isn't necessarily suited to being backed
by our default persistent storage capability (Ceph).

Default to 0 and leave the option to change this under the advanced
section, along with a suitable warning.

Implements CLDNTV-204.