Implicit Child Object Creation (#117)
Server has been modified to implicitly create projects and control
planes, so we don't necessarily need to do it in the UI any more.  The
hard bit in this instance is updaing the cluster view as control planes
were static, and can now appear during the life time of the view (and by
extension disappear possibly).  Tweak it so that a refresh does the
control planes, that will implicitly trigger a cluster update.  Also add
some logic to make the "submit" buttons go inactive on click and show a
spinner, and various other usability tweaks.