Fix Scoping Change
When the project scope changes in the cluster view, there is a race
condition where the list of clusters can get updated before the list of
control planes.  This causes a list operation, which will implictly
create a control plane, based on the old scope, in the new project.  We
need to purge the old control plane before updating anything based on
the new token so it isn't magically auto-provisioned.