Implement Themes (#129)
So the "branding" department want to implement something shiny, so we
need to incorporate glitzy backdrops and translucency all over the
place. This is a good time to roll out SVG images, because they are
smaller and actually scale.  Most pertinently, the existing theme is
white based, and they want black, so no time like the present to
implement user preferences.

Implements #43