PoC Item Details
Items are typically displayed in a grid, therefore for it not to look
shambolic, all things should be sized the same, which limits us to fixed
content.  For other things that are variable, or just irrelevant, we
should be able to diplsay them in a free-form item.  Now we could mess
about with modals, but actually it transpires putting them into the grid
as a full width entry is a way better UX as you can click between items.