Add Variables
Two very annoying things are getting fixed here.  First I never know the
version, we want this to be propagated down from the chart.  Second,
because our dev environment uses RFC1918 addresses for the public
address subnet, we need to avoid these when spinning up clusters, which
is a bit of a mare as twiddling them is time consuming.  Instead we can
just push this into environment specific Helm configuration.