Upgrade Items
These are all a bit free form at the moment, make a header and main
section and allow named slots.  Spearate out the item detail from the
item.  Apply some more eyecatching and structured styling.  Then on top
of all that, siphon off the selection logic and make that common.